Extract EclEpsGridProperties class to separate file
